coinbase_kafka_producer/requirements.txt
# Producer library
kafka-python == 2.2.6
websocket-client == 1.8.0

# Model library